refactor: rename switchTheme file and folder
